segregated scripts
