+++ scripts/constant/constant.ts
@@ -1,6 +1,6 @@
 export const constants = {
   address_zero: "0x0000000000000000000000000000000000000000",
-  eth_addr: "0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE",
+  native_address: "0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE",
   max_value:
     "115792089237316195423570985008687907853269984665640564039457584007913129639935",
 };
