+++ scripts/deployment/deployAvaxConnector.ts
