+++ scripts/deployment/deployCompoundMapping.ts
