+++ scripts/deployment/deployConnectorsFromCmd.ts
