+++ scripts/deployment/deployMappingContract.ts
