+++ scripts/tests/buildDSAv2.ts
@@ -1,8 +1,8 @@
 import { ethers } from "hardhat";
 
-import { addresses } from "./constant/addresses";
-import { abis } from "./constant/abis";
-import { abi } from "../deployements/mainnet/Implementation_m1.sol/InstaImplementationM1.json";
+import { addresses } from "../constant/addresses";
+import { abis } from "../constant/abis";
+import { abi } from "../../deployements/mainnet/Implementation_m1.sol/InstaImplementationM1.json";
 
 export async function buildDSAv2(owner: any) {
   const instaIndex = await ethers.getContractAt(
