+++ scripts/tests/encodeSpells.ts
@@ -1,13 +1,12 @@
 import { web3 } from "hardhat";
-import { addresses } from "./constant/addresses";
-import { abis } from "./constant/abis";
+import { abis } from "../constant/abis";
 
 export function encodeSpells(spells: any[]) {
   const targets = spells.map((a) => a.connector);
   const calldatas = spells.map((a) => {
     const functionName = a.method;
     // console.log(functionName)
-    const abi = abis.connectors[a.connector].find((b) => {
+    const abi = abis.connectors[a.connector].find((b: { name: any }) => {
       return b.name === functionName;
     });
     // console.log(functionName)
