+++ test/mainnet/b.protocol/b.maker.test.ts
@@ -4,10 +4,10 @@ const { web3, deployments, waffle, ethers } = hre;
 const { provider, deployContract } = waffle
 
 
-import { deployAndEnableConnector } from "../../../scripts/deployAndEnableConnector.js"
-import { buildDSAv2 } from "../../../scripts/buildDSAv2"
-import { encodeSpells } from "../../../scripts/encodeSpells.js"
-import { getMasterSigner } from "../../../scripts/getMasterSigner"
+import { deployAndEnableConnector } from "../../../scripts/tests/deployAndEnableConnector.js"
+import { buildDSAv2 } from "../../../scripts/tests/buildDSAv2"
+import { encodeSpells } from "../../../scripts/tests/encodeSpells.js"
+import { getMasterSigner } from "../../../scripts/tests/getMasterSigner"
 import { addresses } from "../../../scripts/constant/addresses";
 import { abis } from "../../../scripts/constant/abis";
 import { constants } from "../../../scripts/constant/constant";
