+++ test/mainnet/basic-ERC721/ERC721-transfer.ts
@@ -5,10 +5,10 @@ const { provider, deployContract } = waffle
 import { abi : implementationsABI } from "../../../scripts/constant/abi/core/InstaImplementations.json"
 
 
-import { deployAndEnableConnector } from "../../../scripts/deployAndEnableConnector.js"
-import { buildDSAv2 } from "../../../scripts/buildDSAv2"
-import { encodeSpells } from "../../../scripts/encodeSpells.js"
-import { getMasterSigner } from "../../../scripts/getMasterSigner"
+import { deployAndEnableConnector } from "../../../scripts/tests/deployAndEnableConnector.js"
+import { buildDSAv2 } from "../../../scripts/tests/buildDSAv2"
+import { encodeSpells } from "../../../scripts/tests/encodeSpells.js"
+import { getMasterSigner } from "../../../scripts/tests/getMasterSigner"
 
 
 import { addresses } from "../../../scripts/constant/addresses"
