+++ test/mainnet/pooltogether-polygon/pooltogether.test.ts
@@ -5,10 +5,10 @@ const { provider, deployContract } = waffle
 
 const ALCHEMY_ID = process.env.ALCHEMY_ID;
 
-import { deployAndEnableConnector } from "../../../scripts/deployAndEnableConnector.js"
-import { buildDSAv2 } from "../../../scripts/buildDSAv2"
-import { encodeSpells } from "../../../scripts/encodeSpells.js"
-import { getMasterSigner } from "../../../scripts/getMasterSigner"
+import { deployAndEnableConnector } from "../../../scripts/tests/deployAndEnableConnector.js"
+import { buildDSAv2 } from "../../../scripts/tests/buildDSAv2"
+import { encodeSpells } from "../../../scripts/tests/encodeSpells.js"
+import { getMasterSigner } from "../../../scripts/tests/getMasterSigner"
 
 
 import { addresses } from "../../../scripts/constant/addresses"
